Make all example Flows visualizable
