+++ docs/content/design-patterns/parallel-subflows/back-pressure.mdx
@@ -16,8 +16,8 @@ This separates admission from processing: parent Flows remain bounded, while sub
 
 <SdkTabs
   examples={{
-    python: 'examples/python/dex_examples/patterns/parallel-subflows/flows.py',
-    go: 'examples/go/patterns/parallel-subflows/flows.go',
+    python: 'examples/python/dex_examples/patterns/parallel-subflows/submit_request_flow.py',
+    go: 'examples/go/patterns/parallel-subflows/submit_request_flow.go',
     java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/SubmitRequestFlow.java',
     typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts',
     rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs',
