+++ docs/content/design-patterns/parallel-subflows/basic.mdx
@@ -12,7 +12,7 @@ Start one SubFlow for each input and wait for every SubFlow to complete.
 
 ## Core implementation
 
-<SdkTabs examples={{python: 'examples/python/dex_examples/patterns/parallel-subflows/flows.py', go: 'examples/go/patterns/parallel-subflows/flows.go', java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/BasicParentFlow.java', typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts', rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs'}}>
+<SdkTabs examples={{python: 'examples/python/dex_examples/patterns/parallel-subflows/basic_parent_flow.py', go: 'examples/go/patterns/parallel-subflows/basic_parent_flow.go', java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/BasicParentFlow.java', typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts', rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs'}}>
 <SdkSnippet lang="python">
 
 ```python
