+++ docs/content/design-patterns/parallel-subflows/wait-for-half.mdx
@@ -14,7 +14,7 @@ The important detail is that the parent does not complete immediately after reac
 
 ## Core implementation
 
-<SdkTabs examples={{python: 'examples/python/dex_examples/patterns/parallel-subflows/flows.py', go: 'examples/go/patterns/parallel-subflows/flows.go', java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/WaitForHalfParentFlow.java', typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts', rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs'}}>
+<SdkTabs examples={{python: 'examples/python/dex_examples/patterns/parallel-subflows/wait_for_half_parent_flow.py', go: 'examples/go/patterns/parallel-subflows/wait_for_half_parent_flow.go', java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/WaitForHalfParentFlow.java', typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts', rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs'}}>
 <SdkSnippet lang="python">
 
 ```python
