+++ docs/content/primitives/subflow.mdx
@@ -18,8 +18,8 @@ The target Flow must be registered on the Worker, and its start Step must accept
 
 <SdkTabs
 examples={{
-  python: 'examples/python/dex_examples/primitives/subflow/subflow_flow.py',
-  go: 'examples/go/primitives/subflow/workflow.go',
+  python: 'examples/python/dex_examples/primitives/subflow/parent_flow.py',
+  go: 'examples/go/primitives/subflow/parent_flow.go',
   java: 'examples/java/src/main/java/io/superdurable/dex/primitives/subflow/SubFlowParentFlow.java',
   typescript: 'examples/typescript/src/primitives/subflow/subflow-flow.ts',
   rust: 'examples/rust/src/primitives/subflow/flow.rs',
