+++ docs/i18n/zh-Hans/docusaurus-plugin-content-docs/current/design-patterns/parallel-subflows/partitioning.mdx
@@ -18,8 +18,8 @@ title: "高级：多父 Flow 分区"
 
 <SdkTabs
   examples={{
-    python: 'examples/python/dex_examples/patterns/parallel-subflows/flows.py',
-    go: 'examples/go/patterns/parallel-subflows/flows.go',
+    python: 'examples/python/dex_examples/patterns/parallel-subflows/submit_request_flow.py',
+    go: 'examples/go/patterns/parallel-subflows/submit_request_flow.go',
     java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/SubmitRequestFlow.java',
     typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts',
     rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs',
@@ -65,9 +65,12 @@ func enqueueRequest(ctx context.Context, client *dex.Client, parentFlow *Advance
 }
 
 func partition(request string, partitions int) int {
-	hash := fnv.New32a()
-	_, _ = hash.Write([]byte(request))
-	return int(hash.Sum32()) % partitions
+	hash := uint32(2_166_136_261)
+	for _, value := range []byte(request) {
+		hash ^= uint32(value)
+		hash *= 16_777_619
+	}
+	return int(hash) % partitions
 }
 ```
