+++ docs/i18n/zh-Hans/docusaurus-plugin-content-docs/current/design-patterns/parallel-subflows/wait-for-half.mdx
@@ -14,7 +14,7 @@ title: 等待一半完成
 
 ## 核心实现
 
-<SdkTabs examples={{python: 'examples/python/dex_examples/patterns/parallel-subflows/flows.py', go: 'examples/go/patterns/parallel-subflows/flows.go', java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/WaitForHalfParentFlow.java', typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts', rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs'}}>
+<SdkTabs examples={{python: 'examples/python/dex_examples/patterns/parallel-subflows/wait_for_half_parent_flow.py', go: 'examples/go/patterns/parallel-subflows/wait_for_half_parent_flow.go', java: 'examples/java/src/main/java/io/superdurable/dex/patterns/parallelsubflows/WaitForHalfParentFlow.java', typescript: 'examples/typescript/src/patterns/parallel-subflows/parallel-subflows.ts', rust: 'examples/rust/src/patterns/parallel_subflows/flow.rs'}}>
 <SdkSnippet lang="python">
 
 ```python
