+++ docs/i18n/zh-Hans/docusaurus-plugin-content-docs/current/primitives/subflow.mdx
@@ -19,8 +19,8 @@ target Flow 必须注册在 Worker 上，而且它的 start Step 必须能接受
 
 <SdkTabs
 examples={{
-  python: 'examples/python/dex_examples/primitives/subflow/subflow_flow.py',
-  go: 'examples/go/primitives/subflow/workflow.go',
+  python: 'examples/python/dex_examples/primitives/subflow/parent_flow.py',
+  go: 'examples/go/primitives/subflow/parent_flow.go',
   java: 'examples/java/src/main/java/io/superdurable/dex/primitives/subflow/SubFlowParentFlow.java',
   typescript: 'examples/typescript/src/primitives/subflow/subflow-flow.ts',
   rust: 'examples/rust/src/primitives/subflow/flow.rs',
