+++ examples/go/patterns/parallel-subflows/basic_parent_flow.go
@@ -1,0 +1,61 @@
+// Copyright (c) 2022-2026 Super Durable, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package parallelsubflows
+
+import "github.com/superdurable/dex/sdk-go/dex"
+
+type BasicParentFlow struct {
+	dex.FlowDefaults
+	exampleFlow *ExampleSubFlow
+}
+
+func NewBasicParentFlow(exampleFlow *ExampleSubFlow) *BasicParentFlow {
+	return &BasicParentFlow{exampleFlow: exampleFlow}
+}
+
+func (flow *BasicParentFlow) GetSteps() []dex.StepDef {
+	return []dex.StepDef{dex.DefineStartStep(subFlowsStep{exampleFlow: flow.exampleFlow})}
+}
+
+func (*BasicParentFlow) GetPersistenceSchema() dex.PersistenceSchema {
+	return dex.PersistenceSchema{}
+}
+
+type subFlowsStep struct {
+	dex.StepDefaults
+	exampleFlow *ExampleSubFlow
+}
+
+func (subFlowsStep) GetStepType() string { return "SubFlowsStep" }
+
+func (step subFlowsStep) WaitFor(_ dex.Context, requests []string) (*dex.Wait, error) {
+	conditions := make([]dex.Condition, 0, len(requests))
+	for _, request := range requests {
+		conditions = append(conditions, dex.SubFlow(step.exampleFlow, request))
+	}
+	return dex.AllOf(conditions...), nil
+}
+
+func (subFlowsStep) Execute(_ dex.Context, _ []string) (*dex.StepDecision, error) {
+	return dex.GracefulComplete(nil), nil
+}
+
+var _ dex.Flow = (*BasicParentFlow)(nil)
