+++ examples/python/dex_examples/patterns/parallel-subflows/models.py
@@ -1,0 +1,30 @@
+# Copyright (c) 2022-2026 Super Durable, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+
+DEFAULT_CONCURRENCY = 10
+MAX_BUFFERED_REQUESTS = 100
+
+
+@dataclass(frozen=True)
+class ParentInput:
+    requests: list[str]
+    concurrency: int = DEFAULT_CONCURRENCY
+
+
+@dataclass(frozen=True)
+class SubmitRequestInput:
+    request: str
+    parent_ids: list[str]
